requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
bcrypt==4.1.2
python-decouple==3.8
mysqlclient==2.2.1
python-dotenv==1.0.0
